fix: add strict PriceData type checks and fix validate() to use parsed output
- Extract price-related types into src/types/price.ts with comprehensive
  JSDoc and isPriceData() runtime type guard
- Update src/types/index.ts to re-export all types from price.ts
- Fix validate() to return result.data (parsed Zod output) on success
  instead of always casting raw data — this ensures schema defaults,
  coercions, and transformations are applied downstream
- Strengthen PriceData and PriceHistoryEntry schemas:
  - assetPair: z.string().min(1)
  - price: z.number().finite()
  - timestamp: z.number().int().min(0)
  - confidence: z.number().min(0).max(1)
  - sources: z.array(z.string().min(1))

Closes #179
