+++ src/types/index.ts
@@ -1,60 +1,23 @@
-export interface PriceData {
-  assetPair: string
-  price: number
-  timestamp: number
-  confidence: number
-  sources: string[]
-}
+// Price-related types — see src/types/price.ts for full documentation
+export type {
+  PriceData,
+  PriceSyncState,
+  LivePriceEntry,
+  PriceHistoryEntry,
+  PriceHistoryResponse,
+  SourceName,
+  SourceHealth,
+  WsSubscribeMessage,
+  WsUnsubscribeMessage,
+  WsPriceUpdate,
+  WsMessage,
+} from './price'
 
-export type PriceSyncState = 'optimistic' | 'confirmed' | 'rollback' | 'synced'
+export { isPriceData } from './price'
 
-export interface LivePriceEntry {
-  data: PriceData
-  syncState: PriceSyncState
-  flashVersion: number
-}
-
-export interface PriceHistoryEntry {
-  price: number
-  timestamp: number
-  confidence: number
-  sources: string[]
-}
-
-export interface PriceHistoryResponse {
-  pair: string
-  history: PriceHistoryEntry[]
-}
-
-export type SourceName = 'chainlink' | 'redstone' | 'band' | 'reflector'
-
-export interface SourceHealth {
-  source: SourceName
-  status: 'healthy' | 'degraded' | 'down'
-  lastUpdate: number | null
-  latency: number | null
-}
-
-export interface WsSubscribeMessage {
-  action: 'subscribe'
-  assetPairs: string[]
-}
-
-export interface WsUnsubscribeMessage {
-  action: 'unsubscribe'
-  assetPairs: string[]
-}
-
-export interface WsPriceUpdate {
-  type: 'price_update'
-  assetPair: string
-  price: number
-  timestamp: number
-  confidence: number
-  sources: string[]
-}
-
-export type WsMessage = WsPriceUpdate
+// ---------------------------------------------------------------------------
+// Alert types
+// ---------------------------------------------------------------------------
 
 export interface Alert {
   id: string
@@ -86,6 +49,10 @@ export interface AlertsContextType {
   togglePanel: () => void
   markAsRead: (id: string) => void
 }
+
+// ---------------------------------------------------------------------------
+// Rate-limit types
+// ---------------------------------------------------------------------------
 
 export interface RateLimitInfo {
   limit: number
